Add test fixtures to testbench scipt.
For each test, the script creates an empty working directory, and passes
to path as the first argument to the test. If the test succeeds, the
directory is removed. If the test fails, the directory is left in place.
With the '--no-delete' argument, the script will preserve the fixture
directory even on success. The path to the fixture directory can be
changed with the '--fixture' argument.