Add tests for finding leaked file descriptors.
The testlib will read all open file descriptors on startup, and make
sure that all file descriptors open on exit were also present at
startup.

This is not foolproof, as it won't notice a leaked descriptor that has
the same number as a descriptor open at start, but this shouldn't happen
unless descriptors are closed on startup.

This is similar to valgrind's --track-fds option, but will not cause
errors if a descriptor open at startup is 'leaked'. Valgrind should be
used for tracking down errors once they are found, but it does not work
with existing test scripts.