Refactor DB tests to work with run-test.pl script.
The new run-test.pl script does a chdir() before running the tests, and
does not pass the fixture path as an argument. The logic from the DB
tests for checking the argument and using it for the fixture directory
is removed. Instead, all tests should use CWD as the fixture.