Fix cleanup of generated docs in Makefile.
The clean target would only clean explicitly-listed documentation pages.
If a man page was generated as a side effect for another page (i.e. two
commands described in the same page) the generated file would not be
removed.

This fix will instead remove all man pages in the doc directory.