Add interface for checking the existence of a database.
The co_exists() function is used to check if a database exists in the
given directory. This check is not performed during co_open() in order
to skip a (possibly unnecessary) system call when opening the database.
It is up to the calling application whether or not to check before
opening.