Use LD_LIBRARY_PATH when testing with cotest.
Previous versions of cotest used LD_PRELOAD to load the library, but
this is incorrect and fails when no globally-available libcobalt.so is
available. The linker needs to resolve all symbols, as well as resolve
all dynamic libraries, which is not satisfied with LD_PRELOAD.

Using LD_LIBRARY_PATH to search for the local library allows overriding
the global library, even in cases where the global library does not
exist.